Add a script for maintaining images on the wiki.
The script does a lot of stuff - it downloads images from both the wiki and
Fallen London, compares them, provides summary reports, and stores everything
in a SQLite DB so that everything is cached and doesn't need to be re-fetched
over and over.

Probably this should be in a separate repo. Oh well.